// client/src/api/models/Category.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Category = {
    id?: string;
    name?: string;
    parent_id?: string | null;
    image_url?: string;
    sku?: string;
    is_active?: boolean;
    created_at?: string;
    updated_at?: string;
};